Server: database file now loaded from config
The database file path will now be loaded from the configuration file if
one is not specified on the command line.
The default database file location is now also ~/.syncro/data.db rather
than syncro.db in the current path.  This should play nicer with MacOS
and linux servers (and not really cause any problems on windows)